Phases 14-15: validation, COUPLING in tune over its range, UI contrast pass, release candidate
Validation (all on a clean rebuild of this commit)
- Release suite: 86 tests / 940 checks, 0 failures. Clean rebuild from an
  empty directory: 4 min 18 s, 0 warnings.
- Realtime: a global operator new hook counts 0 audio-thread allocations
  in 640 blocks with MIDI, preset / gesture changes and structural
  automation; concurrency stress of audio vs message thread.
- Host tests: block sizes 1-4096 bit-identical, sample-rate changes
  22.05-192 kHz, isolated instances, bypass / suspend / editor open-close.
- GUI load: the editor at 60 fps does not slow the audio thread (8.8 vs
  8.9 % of a core); the dirty regions cost 3.0 ms per frame.
- ASan+UBSan full suite and TSan threaded tests: no reports.
- pluginval strictness 10 with GUI: SUCCESS. Steinberg VST3 validator
  (SDK 3.8, built from source): 47/47 standard, 537/537 extensive.
- Standalone launch check under Xvfb.

COUPLING was out of tune at the top of its range (found while measuring
the whole knob): 12-40 cents off at 0.65, up to -186 cents at 1.0. The
compensation's +-12 % authority clamp was reached at 0.62-0.64 on every
material; beyond ~0.75 rad the CORE dissolves into composite modes.
Authority is now +-30 % and the COUPLING curve ends at the in-tune limit
(0.7277 c^0.9432, default 0.35 unchanged); presets were converted to keep
their rotation (audit figures identical), RANDOM windows converted, field
connection weights renormalised. All four materials now stay within
1.2 cents to 0.65 and 6.3 cents at the maximum; BOW and AIR speak over the
whole range. The tuning gate covers the full range and all materials.

UI contrast and legibility pass, measured against the locked reference:
cooler, darker silver tokens, label ink matching the reference, medium
macro labels, larger tile icons and labels, crisp tile outlines, polished
knob chrome with visible value dots. Silver share 66.7 % vs 67.7 %.

Fixes: per-instance editor mouse state (was a function static), node
readouts show a dash before any audio has run (was "x0.00"), explicit
float-to-double conversions in Gesture.cpp, parallel GCC LTO links.

Docs: PRODUCT_SPEC, DSP_ARCHITECTURE, UI_SYSTEM, PARAMETERS, TESTING,
CORE_QUALITY_GATE, RELEASE_CANDIDATE_REPORT, README, development log
Phases 11-15, screenshots and release measurements.

+++ Source/Motion/Gesture.cpp
@@ -87,7 +87,7 @@ Gesture buildGesture (const std::vector<GestureSample>& raw, double beatsPerSeco
             ++k;
         const double ta = raw[k].timeSeconds, tb = raw[k + 1].timeSeconds;
         const double f = tb > ta ? std::clamp ((t - ta) / (tb - ta), 0.0, 1.0) : 0.0;
-        r[static_cast<size_t> (p)] = raw[k].radius + f * (raw[k + 1].radius - raw[k].radius);
+        r[static_cast<size_t> (p)] = static_cast<double> (raw[k].radius) + f * static_cast<double> (raw[k + 1].radius - raw[k].radius);
         a[static_cast<size_t> (p)] = ang[k] + f * (ang[k + 1] - ang[k]);
     }
 
@@ -157,8 +157,8 @@ Gesture orbitGesture (float seconds, float beats, float turns, float radiusWobbl
     for (int i = 0; i < Gesture::kPoints; ++i)
     {
         const double t = static_cast<double> (i) / Gesture::kPoints;
-        g.dAngle[static_cast<size_t> (i)] = wrapAngle (static_cast<float> (2.0 * kPi * turns * t));
-        g.dRadius[static_cast<size_t> (i)] = radiusWobble * static_cast<float> (std::sin (2.0 * kPi * wobbleCycles * t));
+        g.dAngle[static_cast<size_t> (i)] = wrapAngle (static_cast<float> (2.0 * kPi * static_cast<double> (turns) * t));
+        g.dRadius[static_cast<size_t> (i)] = radiusWobble * static_cast<float> (std::sin (2.0 * kPi * static_cast<double> (wobbleCycles) * t));
     }
     g.durationSeconds = std::max (0.05f, seconds);
     g.durationBeats = std::max (0.0f, beats);
